+++ src/Resources/app/storefront/src/checkout/swag-paypal.acdc-fields.js
@@ -222,6 +222,9 @@ export default class SwagPaypalAcdcFields extends SwagPaypalAbstractButtons {
                 // Country Code
                 countryCodeAlpha2: '',
             },
+            contingencies: [
+                'SCA_ALWAYS',
+            ],
         },
 
         /**