+++ src/RestApi/V2/Api/Order/PaymentSource.php
@@ -12,6 +12,7 @@ use Swag\PayPal\RestApi\PayPalApiStruct;
 use Swag\PayPal\RestApi\V2\Api\Order\PaymentSource\Bancontact;
 use Swag\PayPal\RestApi\V2\Api\Order\PaymentSource\Blik;
 use Swag\PayPal\RestApi\V2\Api\Order\PaymentSource\Boletobancario;
+use Swag\PayPal\RestApi\V2\Api\Order\PaymentSource\Card;
 use Swag\PayPal\RestApi\V2\Api\Order\PaymentSource\Eps;
 use Swag\PayPal\RestApi\V2\Api\Order\PaymentSource\Giropay;
 use Swag\PayPal\RestApi\V2\Api\Order\PaymentSource\Ideal;
@@ -46,7 +47,12 @@ class PaymentSource extends PayPalApiStruct
     /**
      * @OA\Property(ref="#/components/schemas/swag_paypal_v2_order_payment_source_boletobancario")
      */
-    protected ?Boletobancario $Boletobancario = null;
+    protected ?Boletobancario $boletobancario = null;
+
+    /**
+     * @OA\Property(ref="#/components/schemas/swag_paypal_v2_order_payment_source_card")
+     */
+    protected ?Card $card = null;
 
     /**
      * @OA\Property(ref="#/components/schemas/swag_paypal_v2_order_payment_source_eps")
@@ -125,12 +131,22 @@ class PaymentSource extends PayPalApiStruct
 
     public function getBoletobancario(): ?Boletobancario
     {
-        return $this->Boletobancario;
+        return $this->boletobancario;
     }
 
-    public function setBoletobancario(?Boletobancario $Boletobancario): void
+    public function setBoletobancario(?Boletobancario $boletobancario): void
     {
-        $this->Boletobancario = $Boletobancario;
+        $this->boletobancario = $boletobancario;
+    }
+
+    public function getCard(): ?Card
+    {
+        return $this->card;
+    }
+
+    public function setCard(?Card $card): void
+    {
+        $this->card = $card;
     }
 
     public function getEps(): ?Eps