+++ tests/Mock/PayPal/Client/_fixtures/V2/GetOrderCapture.php
@@ -71,6 +71,20 @@ class GetOrderCapture
                     ],
                 ],
             ],
+            'payment_source' => [
+                'card' => [
+                    'last_digits' => '7109',
+                    'brand' => 'VISA',
+                    'type' => 'CREDIT',
+                    'authentication_result' => [
+                        'liability_shift' => 'POSSIBLE',
+                        'three_d_secure' => [
+                            'enrollment_status' => 'Y',
+                            'authentication_status' => 'Y',
+                        ],
+                    ],
+                ],
+            ],
             'payer' => [
                 'name' => [
                     'given_name' => self::PAYER_NAME_GIVEN_NAME,